fix codestyle, put qsort in separate module
