+++ task2/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.0.0)
+project(task2)
+
+set(CMAKE_C_STANDARD 99)
+
+add_executable(task2 main.c quickSort.c)
