+++ task2/quickSort.h
@@ -1,0 +1,9 @@
+#pragma once
+
+void swap(int *firstElement, int *secondElement);
+
+int partition(int *array, int lowIndex, int highIndex);
+
+void insertSort(int *array, int lowIndex, int highIndex);
+
+void quickSort(int *array, int lowIndex, int highIndex);
